# swagger/ft.yaml
openapi: 3.0.0
info:
  title: ftx
  version: '1.0'
servers:
  - url: 'http://localhost:3000'
paths:
...
  '/api/preuser/invite/list':
    get:
      tags: [ preuser ]
      summary: Send invite code to a list of preusers
      operationId: get-preuser-invite-list
      description: Send invite code to a list of preusers
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/PreusersIdListSuccesfullySent'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreusersInviteList'
        description: Subset of UserProfile which regular non-privileged user is allowed to update
      security:
        - API Key: []
...
  '/api/playlist':
    put:
      tags: [ playlist ]
      summary: Playlist creating - admin only
      operationId: put-api-playlist
      description: create new playlist
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/PlaylistCreation'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: []
  '/api/playlist/{playlistId}':
    get:
      tags: [ playlist ]
      summary: get playlist by Id - admin only
      operationId: get-api-playlist-{playlistId}
      description: retrieves playlist by Id
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "playlistId"
          in: 'path'
          description: "playlistId"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Playlist'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: []
    post:
      tags: [ playlist ]
      summary: update playlist by Id - admin only
      operationId: post-api-playlist-{playlistId}
      description: update existing playlist by Id
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "playlistId"
          in: 'path'
          description: "playlistId"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/PlaylistUpdation'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: []
  '/api/playlist/{playlistId}/thumbnail':
    post:
      tags: [ playlist ]
      summary: upload thumbnail for the playlist - admin only
      operationId: get-api-playlist-{playlistId}
      description: upload thumbnail for the playlist by Id
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "playlistId"
          in: 'path'
          description: "playlistId"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Playlist'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: []
  '/api/playlist/search':
    get:
      tags: [ playlist ]
      summary: Playlist searching - admin only
      operationId: get-playlist-search
      description: searching for playlist
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "q"
          in: 'query'
          description: "Query string for searching"
          schema:
            type: "string"
        - name: "tags"
          in: 'query'
          description: "tags that describes playlists"
          schema:
            type: array
            items:
              type: string
        - name: "sub"
          in: 'query'
          description: "flag that expands playlist nested objects"
          schema:
            type: "boolean"
        - name: 'pageSize'
          in: 'query'
          description: "maximum number of items per page after the search"
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Playlist'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: []
  '/api/playlist/search/all':
    get:
      tags: [ playlist ]
      summary: Playlist searching all - admin only
      operationId: get-playlist-search-all
      description: searching for playlist all
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "q"
          in: 'query'
          description: "Query string for searching"
          schema:
            type: "string"
        - name: "tags"
          in: 'query'
          description: "tags that describes playlists"
          schema:
            type: array
            items:
              type: string
        - name: "sub"
          in: 'query'
          description: "flag that expands playlist nested objects"
          schema:
            type: "boolean"
        - name: 'pageSize'
          in: 'query'
          description: "maximum number of items per page after the search"
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Playlist'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: []
...
  '/api/system-config/list/{groupKey}/{language}':
    get:
      summary: Gets a list of configs using groupKey and language
      description: ''
      operationId: get-list-api-system-config
      parameters:
        - name: AuthToken
          in: header
          description: Authorization token
          required: true
          schema:
            type: string
        - name: groupKey
          in: path
          description: pitchId
          required: true
          schema:
            type: string
        - name: language
          in: path
          description: pitchId
          required: true
          schema:
            type: string
      responses:
          '200':
            description: OK
            content:
              application/json:
                schema:
                  oneOf:
                    - $ref: '#/components/schemas/SystemConfigsResponse'
                    - $ref: '#/components/schemas/GenericErrors'
          '401':
            description: Unauthorized
          '403':
            description: Forbidden
          '500':
            description: Internal Server Error

      tags:
        - System Config
...
components:
...
  schemas:
...
    Playlist:
      properties:
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        public:
          type: boolean
          default: false
        previewImageUrl:
          type: string
        videos:
          type: array
          items:
            $ref: '#/components/schemas/Video'
        autoPlay:
          type: boolean
          default: false
        playlistAuthor:
          type: object
          $ref: '#/components/schemas/UserProfileUserEditable'
        companyPlaylist:
          type: boolean
          default: false
...
    Video:
      properties:
        title:
          type: string
        description:
          type: string
        previewImageUrl:
          type: string
        tags:
          type: string
        authors:
          type: array
          items:
            $ref: '#/components/schemas/Author'
        likes:
          type: array
          items:
            $ref: '#/components/schemas/UserProfileUserEditable'
        likesCount:
          type: number
        accessGroup:
          type: array
          items:
            type: string
            enum:
              - 'PREMIUM'
    Author:
      properties:
        name:
          type: string
        surname:
          type: string
        title:
          type: string
        description:
          type: string
        avatarUrl:
          type: string
...
    PlaylistCreation:
      description: 'Create new playlist'
      type: object
      x-examples:
        example-1:
          title: string
          description: string
          tags: array
          public: boolean
          previewImageUrl: string
          videos: array
          autoPlay: boolean
          playlistAuthor: string
          companyPlaylist: boolean
      title: PlaylistCreation
      properties:
        title:
          type: string
          description: playlist's title
          example: english grammar
        description:
          type: string
          description: playlist's description
          example: learn english grammar from the very start
        tags:
          type: array
          description: tags that describes the playlist topic
          example: ['english','grammar']
        public:
          type: boolean
          description: is playlist public
          default: false
          example: true
        previewImageUrl:
          type: string
          description: url to preview image
          example: https://previesimages.com/preview/341
        videos:
          type: array
          description: video id array
          example: ['961d61a214c6fd937c1ac3c2', '761da1a214c6fdc37c1ac3c4']
        autoPlay:
          type: boolean
          description: videos in playlist should start play automatically
          default: false
          example: true
        playlistAuthor:
          type: string
          description: author Id
          example: 721d61a214c6fd937c1ac314
        companyPlaylist:
          type: boolean
          description: playlist belongs to a company
          default: false
          example: true
    PlaylistUpdation:
      description: 'Update existing playlist by Id'
      type: object
      x-examples:
        example-1:
          title: string
          description: string
          tags: array
          public: boolean
          previewImageUrl: string
          videos: array
          autoPlay: boolean
          playlistAuthor: string
          companyPlaylist: boolean
      title: PlaylistUpdation
      properties:
        title:
          type: string
          description: updated playlist's title
          example: english grammar updated edition
        description:
          type: string
          description: updated playlist's description
          example: how to learn english grammar way faster
        tags:
          type: array
          description: updated tags that describes the playlist topic
          example: ['english','grammar', 'fast learning']
        public:
          type: boolean
          description: is playlist public
          default: false
          example: true
        previewImageUrl:
          type: string
          description: new url to preview image
          example: https://previesimages.com/preview/741
        videos:
          type: array
          description: video id updated array
          example: ['811d61a214c6fd342c1ac3a7']
        autoPlay:
          type: boolean
          description: videos in playlist should start play automatically
          default: false
          example: true
        playlistAuthor:
          type: string
          description: author Id
          example: 721d61a214c6fd937c1ac314
        companyPlaylist:
          type: boolean
          description: playlist belongs to a company
          default: false
          example: true
...
    PreusersInviteList:
      description: 'Array of preuser ids'
      type: array
      example: [
          "5fd61c19631e2d86c5ae9cf0",
          "6fd61c19631e2d86c5ae9cf1",
          "7fd61c19631e2d86c5ae9cf2",
      ]
    PreusersIdListSuccesfullySent:
      description: 'Array of preuser ids that invite code has been succesfully sent'
      type: array
      example: [
          "5fd61c19631e2d86c5ae9cf0",
          "6fd61c19631e2d86c5ae9cf1",
          "7fd61c19631e2d86c5ae9cf2",
      ]
...
    UserProfileUserEditable:
      description: ''
      type: object
      x-examples:
        example-1:
          email: string
          name: string
          surname: string
          avatarUrl: string
          bio: string
          theme: string
          watchedVideos: 0
      title: UserProfileUserEditable
      properties:
        email:
          type: string
          minLength: 1
          description: User email. Should be unique across the users
          example: test@domain.com
        name:
          type: string
          minLength: 1
          description: User first name
          example: BoJack
        surname:
          type: string
          minLength: 1
          description: User last name
          example: Horseman
        avatarUrl:
          type: string
          minLength: 1
          description: User's avatar URL
          example: 'https://i.pravatar.cc/300'
        bio:
          type: string
          minLength: 1
          description: User's bio
          example: User's bio goes here
        theme:
          type: string
          minLength: 1
          description: User's client application theme
          enum:
            - light
            - dark
...
    GenericErrors:
      title: GenericErrors
      type: object
      x-examples:
        example-1:
          - string
      properties:
        errors:
          type: array
          description: Errors list occurred during request processing
          items:
            type: string
      description: ''
...
    SystemConfigResponse:
      description: ''
      type: object
      title: SystemConfigResposnse
      properties:
        groupKey:
          type: string
          minLength: 1
          description: Group key for defining lists of configs
          example: 'language'
        itemKey:
          type: string
          minLength: 1
          description: Item key used to define a single config
          example: 'language:english'
        language:
          type: string
          minLength: 1
          description: Locale of the config's value
          example: 'en'
        value:
          type: string
          minLength: 1
          description: The value of the config
          example: 'English'
    SystemConfigsResponse:
      description: ''
      type: object
      title: SystemConfigsResposnse
      properties:
        payload:
          type: array
          items:
            - $ref: '#/components/schemas/SystemConfigResponse'
          example: |-
            [{
                "groupKey": "language",
                "itemKey": "language:english",
                "language": "en",
                "value": "English"
            },
            {
                "groupKey": "language",
                "itemKey": "language:afrikaans",
                "language": "en",
                "value": "Afrikaans"
            },
            {
                "groupKey": "language",
                "itemKey": "language:french",
                "language": "en",
                "value": "French"
            }]
...
  securitySchemes:
    API Key:
      name: API Key
      type: apiKey
      in: header
      description: Firebase token